Feature/stored block index alt 473 (#663)
* Refactoring for using StoredBlockIndex

* Cleanup after merge

* Add PopState to StoredBlockIndex

* Fix linux build

* Fix linux build

* Fix fuzzing

* Fix C interface

* Fix warnings

* Fix C storage

* Use StoredBlockIndex in BlockBatch interface

* Fix build

* Small fix

* Validate stored endorsements

* Cleanup

+++ include/veriblock/pop/entities/merkle_tree.hpp
@@ -153,7 +153,7 @@ struct VbkMerkleTree : public MerkleTree<VbkMerkleTree, uint256> {
 
     VbkMerklePath merklePath;
     merklePath.treeIndex = treeIndex;
-    merklePath.index = index;
+    merklePath.index = (int32_t)index;
     merklePath.subject = hash;
     merklePath.layers = getMerklePathLayers(index);
     return merklePath;
@@ -191,7 +191,7 @@ struct BtcMerkleTree : public MerkleTree<BtcMerkleTree, uint256> {
     size_t index = it->second;
 
     MerklePath merklePath;
-    merklePath.index = index;
+    merklePath.index = (int32_t)index;
     merklePath.subject = hash;
     merklePath.layers = getMerklePathLayers(index);
     return merklePath;
